+++ quran/init.go
@@ -2,7 +2,6 @@ package quran
 
 import (
 	"encoding/xml"
-	"errors"
 	"fmt"
 	"io/ioutil"
 	"os"
@@ -10,31 +9,6 @@ import (
 	"sync"
 	"time"
 )
-
-type Quran struct {
-	Suras []struct {
-		Index int    `xml:"index,attr"`
-		Name  string `xml:"name,attr"`
-		Ayas  []struct {
-			Index     int    `xml:"index,attr"`
-			Text      string `xml:"text,attr"`
-			Bismillah string `xml:"bismillah,attr"`
-		} `xml:"aya"`
-	} `xml:"sura"`
-	root *Node
-}
-
-type Node struct {
-	Locations []Location
-	Children  []Child
-}
-
-type Location struct{ Sura, Aya, SliceIndex int }
-
-type Child struct {
-	Key   rune
-	Value *Node
-}
 
 type Transliteration struct {
 	Hijaiyas map[string][]string
@@ -70,61 +44,6 @@ func getCorpusPath() string {
 		return path
 	}
 	return "corpus/"
-}
-
-// Get aya text from sura number and aya number (number starting from 1)
-func (q Quran) GetAya(sura, aya int) (string, error) {
-	if !(1 <= sura && sura <= len(q.Suras)) {
-		return "", errors.New(fmt.Sprintf("invalid sura number %d", sura))
-	}
-	ayas := q.Suras[sura-1].Ayas
-	if !(1 <= aya && aya <= len(ayas)) {
-		return "", errors.New(fmt.Sprintf("invalid sura number %d and aya number %d", sura, aya))
-	}
-	return ayas[aya-1].Text, nil
-}
-
-// Build index of Quran q
-func (q *Quran) BuildIndex() {
-	q.root = &Node{Locations: emptyLocations}
-	for s, sura := range QuranClean.Suras {
-		for a, aya := range sura.Ayas {
-			indexAya([]rune(aya.Text), s, a, q.root)
-		}
-	}
-}
-
-func indexAya(harfs []rune, sura, aya int, node *Node) {
-	sliceIndex := 0
-	for i := range harfs {
-		if i == 0 || harfs[i-1] == ' ' {
-			buildTree(harfs[i:], Location{sura, aya, sliceIndex}, node)
-			sliceIndex++
-		}
-	}
-}
-
-func buildTree(harfs []rune, location Location, node *Node) {
-	for i, harf := range harfs {
-		child := getChild(node.Children, harf)
-		if child == nil {
-			child = &Node{}
-			node.Children = append(node.Children, Child{harf, child})
-		}
-		node = child
-		if i == len(harfs)-1 || harfs[i+1] == ' ' {
-			node.Locations = append(node.Locations, location)
-		}
-	}
-}
-
-func getChild(children []Child, key rune) *Node {
-	for _, child := range children {
-		if child.Key == key {
-			return child.Value
-		}
-	}
-	return nil
 }
 
 func loadTransliterationAsync(wg *sync.WaitGroup, fileName string, t *Transliteration) {
